Creates AttitudeEstimator class for attitude estimation
Creates the AttitudeEstimator class. It represents a set of
attitude estimations through a vector in which each element is an
AttitudeEstimation struct. The AttitudeEstimation struct
represents a single attitude estimation composed of the timestamp
of the corresponding measurements and the estimated roll and pitch
angles. These angles are estimated in the AttitudeEstimator class
by following the aerospace rotation sequence.